Add unit test create.sql, UnitTests
Fix Country Update sql
Fix Station Rowmapper conversion exception (wrong datatype at rowmapper and dataclass)

+++ Wetr-Solution/Wetr.Domain/Station.cs
@@ -9,8 +9,8 @@ namespace Wetr.Domain
         public int AddressId { get; set; }
         public int UserId { get; set; }
         public string Name { get; set; }
-        public double Longitude { get; set; }
-        public double Latitude { get; set; }
+        public decimal Longitude { get; set; }
+        public decimal Latitude { get; set; }
 
         public override string ToString() =>
             $"[{StationId}] {StationTypeId} {AddressId} {UserId} {Name} {Longitude} {Latitude}";
